Removes as much kubeclient Get methods as possible
Kubeclient get for nodes and pods were being used in other places in the
code. Removed all of their uses except for specific cases like the ovn
db manager and windows, where we do not have full informer setups. While
transitioning to use the factory, it created a cylical dependency
between metrics and factory libraries, due to the configuration duration
recorder.

Split the configuration duration recorder into its own sub-package under
metrics/recorders.
