+++ go-controller/pkg/metrics/metrics.go
@@ -27,22 +27,11 @@ import (
 	libovsdbclient "github.com/ovn-org/libovsdb/client"
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/types"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
 )
 
 const (
-	MetricOvnkubeNamespace               = "ovnkube"
-	MetricOvnkubeSubsystemController     = "controller"
-	MetricOvnkubeSubsystemClusterManager = "clustermanager"
-	MetricOvnkubeSubsystemNode           = "node"
-	MetricOvnNamespace                   = "ovn"
-	MetricOvnSubsystemDB                 = "db"
-	MetricOvnSubsystemNorthd             = "northd"
-	MetricOvnSubsystemController         = "controller"
-	MetricOvsNamespace                   = "ovs"
-	MetricOvsSubsystemVswitchd           = "vswitchd"
-	MetricOvsSubsystemDB                 = "db"
-
 	ovnNorthd     = "ovn-northd"
 	ovnController = "ovn-controller"
 	ovsVswitchd   = "ovs-vswitchd"
@@ -82,7 +71,7 @@ type stopwatchStatistics struct {
 // resource reached the maximum retry limit and will not be retried. This metric doesn't
 // need Subsystem string since it is applicable for both master and node.
 var MetricResourceRetryFailuresCount = prometheus.NewCounter(prometheus.CounterOpts{
-	Namespace: MetricOvnkubeNamespace,
+	Namespace: types.MetricOvnkubeNamespace,
 	Name:      "resource_retry_failures_total",
 	Help:      "The total number of times processing a Kubernetes resource reached the maximum retry limit and was no longer processed",
 })
